Bug 142887 - role progress bar does not support indeterminate state https://bugs.webkit.org/show_bug.cgi?id=142887
Patch by Michael Peechatt <mpeechatt@gmail.com> on 2015-04-07
Reviewed by Chris Fleizach.

Source/WebCore:

Test: platform/mac/accessibility/progressbar-indeterminate.html

* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:]):
When getting min or max value of an indeterminate progress indicator, return 0.
This is so VoiceOver will recognize it as indeterminate.

LayoutTests:

* platform/mac/accessibility/progressbar-indeterminate-expected.txt: Added.
* platform/mac/accessibility/progressbar-indeterminate.html: Added.
Test that checks if 0 is returned for indetermiate progress indicators when
aria min and max values are non zero.

Identifier: 161536@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182520 268f45cc-cd09-0410-ab3c-d52691b4dbfc