[Cocoa] Localizable strings are inconsistent and need to be regenerated https://bugs.webkit.org/show_bug.cgi?id=143661 Source/WebCore:
rdar://problem/19634388

Reviewed by Dan Bernstein.

* English.lproj/Localizable.strings: Updated by running the
update-webkit-localizable-strings script.

Source/WebKit/mac:

rdar://problem/19634388

Reviewed by Dan Bernstein.

* Misc/WebLocalizableStrings.mm:
(LoadWebLocalizedStringsTimerCallback): Moved the comment, useful for programmers but
not for localization, into a C++ source code comment, and made the localizer comment
consistent with the place this localizable string is used for actual localization.
* WebCoreSupport/WebEditorClient.mm:
(undoNameForEditAction): Changed the localizer comments for Delete and Dictation
to match the ones in WebKit2; no need to specify that these are iOS only, and if we do
want to specify that, we should not be using the term PLATFORM(IOS) since there is no
reason for us to expect localizers to understand that terminology. Also added a comment
about the fact that we have two copies of all the localizable strings.

Source/WebKit2:

Reviewed by Dan Bernstein.

* UIProcess/WebEditCommandProxy.cpp:
(WebKit::WebEditCommandProxy::nameForEditAction): Also added a comment
about the fact that we have two copies of all the localizable strings.

Identifier: 161689@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182735 268f45cc-cd09-0410-ab3c-d52691b4dbfc