We should dump GraphicsLayer's anchorPoint z component https://bugs.webkit.org/show_bug.cgi?id=143815
Reviewed by Tim Horton.

We didn't include the z component of a layer's anchor point when dumping.
Dump if it's non-zero (to avoid having to change lots of test output).
No test with non-zero z appears to dump layers.

* platform/graphics/GraphicsLayer.cpp:
(WebCore::GraphicsLayer::dumpProperties):
* rendering/style/RenderStyle.cpp:
(WebCore::requireTransformOrigin): Remove a FIXME which, on further consideration,
is wrong.


Identifier: 161821@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182880 268f45cc-cd09-0410-ab3c-d52691b4dbfc