Sites with both width=device-width and height=device-height load zoomed out https://bugs.webkit.org/show_bug.cgi?id=143795 <rdar://problem/20369671>
Reviewed by Ben Poulain.

* page/ViewportConfiguration.cpp:
(WebCore::ViewportConfiguration::shouldIgnoreVerticalScalingConstraints):
Some sites specify both width=device-width and height=device-height, and
then lay out to device width but with a large amount of vertically scrollable content
(so, height=device-height was a lie).

In all other cases where we use device-width and device-height, we prefer
width=device-width over height=device-height, but in the code to ignore scaling constraints,
the two paths were completely separate. On sites that specify both, this
resulted in us attempting to zoom out to fit the entire height of the very tall page,
which isn't at all what we wanted. So, ignore height=device-height if a width is specified.


Identifier: 161848@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@182910 268f45cc-cd09-0410-ab3c-d52691b4dbfc