[iOS] When computing visible rects for tiling, stop searching at UIWindows https://bugs.webkit.org/show_bug.cgi?id=144022 <rdar://problem/18327227>
Reviewed by Simon Fraser.

[WAKWindow _visibleRectRespectingMasksToBounds:] computes a visible rect which we use
to determine which tiles to create. We do this by finding the frame of the _hostLayer,
and then walking up the CALayer hierarchy converting each rect into its parent's
coordinate system (all the while clipping to bounds if necessary). This walk up the
layer hierarchy should stop at a layer associated with a UIWindow.

* platform/ios/wak/WAKWindow.mm:
(-[WAKWindow _visibleRectRespectingMasksToBounds:]):

Identifier: 161990@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183089 268f45cc-cd09-0410-ab3c-d52691b4dbfc