Web Inspector: Fix an exception preventing switching timelines https://bugs.webkit.org/show_bug.cgi?id=144053
Reviewed by Darin Adler.

* UserInterface/Views/TreeOutline.js:
(WebInspector.TreeElement.prototype.select): Store treeOutline in a local so
we have a reference to it when we go to reset processingSelectionChange.
The calls to onselect was removing the tree element from the outline, causing
"delete this.treeOutline.processingSelectionChange" to throw an exception.
This patch changes the delete to a set to false for good measure too.


Identifier: 162016@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183119 268f45cc-cd09-0410-ab3c-d52691b4dbfc